# mini-xrm-backend/src/gen/java/com/example/minixrm/backend/web/openapi/v1/api/ReportApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.minixrm.backend.web.openapi.v1.api;

import com.example.minixrm.backend.web.openapi.v1.model.ErrorResponseView;
import com.example.minixrm.backend.web.openapi.v1.model.PersonResponsibleReportPageView;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.19.0")
@Validated
@Controller
public interface ReportApi {

    String PATH_REPORT_PERSON_RESPONSIBLE = "/v1/reports/person-responsible";
    /**
     * GET /v1/reports/person-responsible : Retrieve a report of people responsible for activities
     *
     * @param page The page number to retrieve (0-based) (required)
     * @param pageSize The number of items per page (required)
     * @return Successful response (status code 200)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ReportApi.PATH_REPORT_PERSON_RESPONSIBLE,
        produces = { "application/json" }
    )
    ResponseEntity<PersonResponsibleReportPageView> reportPersonResponsible(
        @NotNull @Min(value = 0)  @Valid @RequestParam(value = "page", required = true) Integer page,
        @NotNull @Min(value = 1)  @Valid @RequestParam(value = "pageSize", required = true) Integer pageSize
    );

}