# mini-xrm-backend/src/main/resources/static/openapi-v1.yaml
openapi: 3.0.1
info:
    title: Mini xRM API
    version: "0.0.1"
paths:
...
    /v1/reports/person-responsible:
        get:
            tags:
                - Report
            operationId: reportPersonResponsible
            parameters:
                - $ref: '#/components/parameters/page'
                - $ref: '#/components/parameters/pageSize'
            summary: Retrieve a report of people responsible for activities
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PersonResponsibleReportPageView'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalServerError'
components:
    parameters:
...
        page:
            name: page
            in: query
            required: true
            schema:
                type: integer
                format: int32
                minimum: 0
                example: 0
            description: The page number to retrieve (0-based)
        pageSize:
            name: pageSize
            in: query
            required: true
            schema:
                type: integer
                format: int32
                minimum: 1
                example: 20
            description: The number of items per page
    responses:
        BadRequest:
            description: Bad Request, e.g. validation failed for the request body or parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponseView'
...
        InternalServerError:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponseView'
...
    schemas:
...
        PageView:
            type: object
            properties:
                totalElements:
                    type: integer
                    format: int64
                    minimum: 0
                    example: 100
                totalPages:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 5
                currentPage:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 0
                    description: The current page number (0-based)
                pageSize:
                    type: integer
                    format: int32
                    minimum: 1
                    example: 20
            required:
                - totalElements
                - totalPages
                - currentPage
                - pageSize
...
        PersonResponsibleReportPageView:
            allOf:
                - $ref: '#/components/schemas/PageView'
                - type : object
                  properties:
                      content:
                          type: array
                          items:
                              $ref: '#/components/schemas/PersonResponsibleReportItemView'
            required:
                - content
        PersonResponsibleReportItemView:
            type: object
            properties:
                personResponsible:
                    type: string
                    minLength: 0
                    maxLength: 150
                    example: "Jane Smith"
                totalDurationMinutes:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 240
                partnerCount:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 5
            required:
                - personResponsible
                - totalDurationMinutes
                - partnerCount
        ErrorResponseView:
            type: object
            properties:
                code:
                    type: string
                    example: VALIDATION_FAILED
                message:
                    type: string
                details:
                    type: string
            required:
                - code
                - message
                - details
...